% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASRtools-package.R
\docType{package}
\name{ASRtools-package}
\alias{ASRtools}
\alias{ASRtools-package}
\title{ASRtools: Extra tools for ASRpackages}
\description{
Complementary functions for ASRpackages to help preparing data or analyses, and to provide with additional tools to explore outputs. Functions starting with dot (.) are not intended to be directly called by the user.
}
\author{
\strong{Maintainer}: Salvador Gezan \email{salvador.gezan@vsni.co.uk}

Authors:
\itemize{
  \item Giovanni Galli \email{giovanni.galli@vsni.co.uk}
  \item Darren Murray \email{darren.murray@vsni.co.uk}
}

Other contributors:
\itemize{
  \item VSN International [copyright holder]
}

}
\keyword{internal}
