% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_status.R
\name{.async_update_status}
\alias{.async_update_status}
\title{Same as \code{.update_status} but asynchronously}
\usage{
.async_update_status(
  progress = NULL,
  context = NULL,
  category = NULL,
  verbose = FALSE
)
}
\arguments{
\item{progress}{An integer representing the progress of the process
(default = \code{NULL}).}

\item{context}{A character string describing the context of the process
(default = \code{NULL}).}

\item{category}{A character string describing the category of the process
(default = \code{NULL}).}

\item{verbose}{A logical value indicating whether to print
information (default = \code{FALSE}).}
}
\description{
This function sends a POST request to the endpoint specified in the
environment variable ENDPOINT. The request body is a JSON object with the
progress, context, and category. The request is sent asynchronously using
a system call to \code{curl}.
}
\examples{
\dontrun{
Sys.setenv(ENDPOINT = "http://localhost:3000/...")
.async_update_status(10L, "loading data", "inProgress", verbose = TRUE)
}
}
