% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blup.R
\name{.blup}
\alias{.blup}
\title{Gets EBLUPs from an \code{asreml} object.}
\usage{
.blup(object, terms = NULL, tidy = TRUE)
}
\arguments{
\item{object}{An \code{asreml} object (default = \code{NULL}).}

\item{terms}{A character vector with the names of the terms to be extracted
from the \code{object} (default = \code{NULL}).}

\item{tidy}{If \code{TRUE}, the \code{terms} are presented in the style
of \code{predict.asreml} (default = \code{TRUE}).}
}
\value{
A list of class \code{blup.asreml} with the EBLUPs from
\code{object}.
}
\description{
Gets EBLUPs from an \code{asreml} object.
}
\details{
This function will work for simple random terms. It does not currently
supports \code{at()} and other constructors from \code{asreml}.
}
\examples{
\dontrun{
library(asreml)

oats.asr <- asreml(yield ~ Variety*Nitrogen,
  random = ~ Blocks/Wplots, data = oats)

.blup(oats.asr, terms = c("Blocks:Wplots"))

}
}
