% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_map.R
\name{.dummy_map}
\alias{.dummy_map}
\title{Creates a dummy genomic map}
\usage{
.dummy_map(marker_id = NULL, message = TRUE)
}
\arguments{
\item{marker_id}{vector with names of markers to compose dummy map.}

\item{message}{logical value indicating whether diagnostic messages should be printed on screen (default = \code{TRUE}).}
}
\value{
Data frame with dummy map. A single chromosome/linkage group is created and marker
distances are a sequence from one to the number of markers.
}
\description{
Creates a dummy genomic map
}
