% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_asreml.R
\name{.try_asreml}
\alias{.try_asreml}
\title{Error handling for ASReml-R}
\usage{
.try_asreml(
  x = "mod",
  call = NULL,
  updates = 2,
  env = parent.frame(),
  message = TRUE,
  ...
)
}
\arguments{
\item{call}{The asreml call to be evaluated provided as character.}

\item{updates}{The maximum number of tries to update the model if it
has not converged yet.}
}
\value{
A list with the exit status and the model object if it converged.
}
\description{
Error handling for ASReml-R
}
