% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_status.R
\name{.update_status}
\alias{.update_status}
\title{Update the status of a process}
\usage{
.update_status(
  progress = NULL,
  context = NULL,
  category = NULL,
  verbose = FALSE
)
}
\arguments{
\item{progress}{An integer representing the progress of the process
(default = \code{NULL}).}

\item{context}{A character string describing the context of the process
(default = \code{NULL}).}

\item{category}{A character string describing the category of the process
(default = \code{NULL}).}

\item{verbose}{A logical value indicating whether to print
information (default = \code{FALSE}).}
}
\value{
The response from the server.
}
\description{
This function sends a POST request to the endpoint specified in the
environment variable ENDPOINT. The request body is a JSON object with the
progress, context, and category
}
\details{
A server must be running at the endpoint specified in the
environment variable ENDPOINT for this function to work. The server must
accept POST requests with a JSON body containing the aforementioned keys.
The older httr packages is used instead of the newer httr2 as it has
fewer dependencies and generates smaller Docker images.
}
\examples{
\dontrun{
library(httr)
# Set the endpoint
Sys.setenv(ENDPOINT = "http://localhost:3000/...")
status <- status_update(50L, "Processing data", "Data processing")
}
}
