% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/git.R
\name{setGitProxy}
\alias{getGitProxy}
\alias{gitPull}
\alias{gitStatus}
\alias{rmGitProxy}
\alias{setGitProxy}
\title{git tools}
\usage{
setGitProxy(proxy, port, auto = "ufrrj")

getGitProxy()

rmGitProxy()

gitStatus()

gitPull()
}
\arguments{
\item{proxy}{Proxy address without the \code{"http://"}.}

\item{port}{Port number.}

\item{auto}{Use a default http.proxy.}
}
\description{
Tools to make the use of git easier.
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}

