% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ASRtools.R
\docType{package}
\name{ASRtools-package}
\alias{ASRtools}
\alias{ASRtools-package}
\title{ASRtools: Tools to make my life easier}
\description{
This is a collection of tools developed for my personal use with the goal of
making my life easier.
}
\details{
\tabular{ll}{
Package: \tab ASRtools  \cr
Type:    \tab Package   \cr
Version: \tab 0.1.1     \cr
Date:    \tab 2015-01-13\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Author and Maintainer: Alessandro Samuel-Rosa
\email{alessandrosamuelrosa@gmail.com}
}

