% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/git.R
\name{setGitProxy}
\alias{getGitProxy}
\alias{rmGitProxy}
\alias{setGitProxy}
\title{git settings}
\usage{
setGitProxy(proxy, port, auto = "ufrrj")

getGitProxy()

rmGitProxy()
}
\arguments{
\item{proxy}{Proxy address without the \code{"http://"}.}

\item{port}{Port number.}

\item{auto}{Use a default http.proxy.}
}
\description{
Tools to configure git.
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}

